requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
ConfigArgParse==0.13.0
cycler==0.10.0
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.15.4
packaging==18.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
urllib3==1.24.1
